move errors.ts to folder
